Implement PUSH and POP. (closed #76 and #77) PUSH stackname thing
command.  Adds the "thing" to the stack that is the value of the
variable whose name is "stackname".  This variable must have a list
as its value; the initial value should be the empty list.  New
members are added at the front of the list.

POP stackname

outputs the most recently PUSHed member of the stack that is the
value of the variable whose name is "stackname" and removes that
member from the stack.